move more script things
