+++ src/private.rs
@@ -9,6 +9,10 @@ use crate::types::{
     layer::Layer,
     level_data::LevelData,
     local_level::LocalLevel,
+    novascript::{
+        action::Action, activator::Activator, nova_value::NovaValue, parameter::Parameter,
+        variable::Variable, NovaScript,
+    },
     object::Object,
     object_property::ObjectProperty,
     pattern::Pattern,
@@ -57,20 +61,20 @@ impl_sealed!(
     BrushObject,
     BrushGrid,
     super::Script,
-    super::NovaScript,
+    NovaScript,
     super::OldAction,
     super::OldActionType,
     super::OldActionProperty,
-    super::Action,
+    Action,
     ActionType,
-    super::NovaValue,
+    NovaValue,
     DynamicType,
     super::FunctionCall,
     super::CallParameter,
-    super::Variable,
+    Variable,
     super::StaticType,
-    super::Activator,
-    super::Parameter,
+    Activator,
+    Parameter,
     Uuid,
 );
